details quincaillerie

+++ src/main/java/com/ict300/P04/Service/favorite/FavoriteService.java
@@ -1,12 +1,14 @@
 package com.ict300.P04.Service.favorite;
 
+import com.ict300.P04.DTO.favorite.request.AddFavoriteProductDTO;
 import com.ict300.P04.DTO.favorite.request.AddFavoriteQuincaillerieDTO;
+import com.ict300.P04.DTO.favorite.request.DeleteFavoriteProductDTO;
 import com.ict300.P04.DTO.favorite.request.DeleteFavoriteQuincaillerieDTO;
-import com.ict300.P04.Entite.FavoriteQuincaillerie;
-import com.ict300.P04.Entite.Quincaillerie;
-import com.ict300.P04.Entite.User;
+import com.ict300.P04.Entite.*;
 import com.ict300.P04.Utilitaires.GenerateID;
+import com.ict300.P04.repository.interfaces.favoriteProduct.FavoriteProductInterface;
 import com.ict300.P04.repository.interfaces.favoriteQuincaillerie.FavoriteQuincaillerieInterface;
+import com.ict300.P04.repository.interfaces.product.ProductInterface;
 import com.ict300.P04.repository.interfaces.quincaillerie.QuincaillerieInterface;
 import com.ict300.P04.repository.interfaces.user.UserInterface;
 import jakarta.transaction.Transactional;
@@ -21,7 +23,13 @@ public class FavoriteService {
     private FavoriteQuincaillerieInterface favoriteQuincaillerieInterface;
 
     @Autowired
+    private FavoriteProductInterface favoriteProductInterface;
+
+    @Autowired
     private QuincaillerieInterface quincaillerieInterface;
+
+    @Autowired
+    private ProductInterface productInterface;
 
     @Autowired
     private UserInterface userInterface;
@@ -29,7 +37,7 @@ public class FavoriteService {
     public void addFavoriteQuincaillerie(AddFavoriteQuincaillerieDTO addFavoriteQuincaillerieDTO){
 
         if(favoriteQuincaillerieInterface.existsByUserIdUserAndQuincaillerieIdQuincaillerie(addFavoriteQuincaillerieDTO.getIdUser(), addFavoriteQuincaillerieDTO.getIdQuincaillerie())){
-            throw new RuntimeException("Produit deja en favoris");
+            throw new RuntimeException("Quincaillerie deja en favoris");
         }
 
         User user = userInterface.getUser(addFavoriteQuincaillerieDTO.getIdUser());
@@ -49,6 +57,36 @@ public class FavoriteService {
 
     @Transactional
     public void deleteFavoriteQuincaillerie(DeleteFavoriteQuincaillerieDTO deleteFavoriteQuincaillerieDTO){
+        if(!favoriteQuincaillerieInterface.existsByUserIdUserAndQuincaillerieIdQuincaillerie(deleteFavoriteQuincaillerieDTO.getIdUser() , deleteFavoriteQuincaillerieDTO.getIdQuincaillerie())){
+            throw new RuntimeException("Quincaillerie deja supprimé des favoris");
+        }
         favoriteQuincaillerieInterface.deleteByUserIdUserAndQuincaillerieIdQuincaillerie(deleteFavoriteQuincaillerieDTO.getIdUser(), deleteFavoriteQuincaillerieDTO.getIdQuincaillerie());
+    }
+
+    public void addFavoriteProduct(AddFavoriteProductDTO addFavoriteProductDTO){
+
+        if(favoriteProductInterface.existsByUserIdUserAndProductIdProduct(addFavoriteProductDTO.getIdUser(), addFavoriteProductDTO.getIdProduct())){
+            throw new RuntimeException("Produit deja en favoris");
+        }
+
+        User user = userInterface.getUser(addFavoriteProductDTO.getIdUser());
+        Product product = productInterface.getProduct(addFavoriteProductDTO.getIdProduct());
+
+        FavoriteProduct favoriteProduct = new FavoriteProduct();
+
+        favoriteProduct.setIdFavoriteProduct(GenerateID.GenerateFavoriteProductID());
+        favoriteProduct.setProduct(product);
+        favoriteProduct.setUser(user);
+        favoriteProduct.setDateAdded(LocalDateTime.now());
+
+        favoriteProductInterface.save(favoriteProduct);
+    }
+
+    @Transactional
+    public void deleteFavoriteProduct(DeleteFavoriteProductDTO deleteFavoriteProductDTO){
+        if(!favoriteProductInterface.existsByUserIdUserAndProductIdProduct(deleteFavoriteProductDTO.getIdUser(), deleteFavoriteProductDTO.getIdProduct())){
+            throw new RuntimeException("Produit deja supprimé des favoris");
+        }
+        favoriteProductInterface.deleteByUserIdUserAndProductIdProduct(deleteFavoriteProductDTO.getIdUser(), deleteFavoriteProductDTO.getIdProduct());
     }
 }
